fix(ci): update xtask patterns check for new layout; split live.rs
Template Contracts (cargo xtask patterns) failures:
- actions check read ACTION_SPECS/help/cli from old monolithic paths; point
  it at src/actions/registry.rs, src/mcp/help.rs (generated), and the
  cli/router.rs tree; map api_put/api_delete CLI verbs. (regression from the
  descriptor-table refactor)
- xtask/src/live.rs was 746>700 effective lines (pre-existing on main since
  the live-suite expansion); split run_rest/run_mcp into live/suites.rs
  (746->461 effective), zero behavior change.

cargo xtask patterns now exits 0; 502 tests green; clippy clean.
